page修改成 @ApiModelProperty(value = "当前页码，从1开始") private long current=0; @ApiModelProperty(value = "每页条数") private long size = 20; @ApiModelProperty(value = "数据总条数") private long total; @ApiModelProperty(value = "数据集合") private List<T> records;

+++ dlz.framework.db/dlz.framework.db.springjdbc/src/main/java/com/dlz/framework/db/helper/support/dbs/DbOpMysql.java
@@ -38,7 +38,7 @@ public class DbOpMysql extends SqlHelper {
      * @return
      */
     @Override
-    public String getLimitSql(int currPage, int pageSize) {
+    public String getLimitSql(long currPage, long pageSize) {
         return " LIMIT " + (currPage - 1) * pageSize + "," + pageSize;
     }
