+++ dlz.framework.db/dlz.framework.db.springjdbc/src/main/java/com/dlz/framework/db/helper/support/dbs/DbOpSqlite.java
@@ -22,7 +22,7 @@ public class DbOpSqlite extends SqlHelper {
     }
 
     @Override
-    public String getLimitSql(int currPage, int pageSize) {
+    public String getLimitSql(long currPage, long pageSize) {
         return " LIMIT " + (currPage - 1) * pageSize + "," + pageSize;
     }
